dns: move message parsing and composition to own module
