+++ sys/Makefile.dep
@@ -106,6 +106,10 @@ ifneq (,$(filter dhcpv6_client,$(USEMODULE)))
   endif
 endif
 
+ifneq (,$(filter dns_%,$(USEMODULE)))
+  USEMODULE += dns
+endif
+
 ifneq (,$(filter fuzzing,$(USEMODULE)))
   USEMODULE += netdev_test
   USEMODULE += gnrc_netif
@@ -470,6 +474,7 @@ ifneq (,$(filter sock_async,$(USEMODULE)))
 endif
 
 ifneq (,$(filter sock_dns,$(USEMODULE)))
+  USEMODULE += dns_msg
   USEMODULE += sock_udp
   USEMODULE += sock_util
   USEMODULE += posix_headers
