+++ sys/net/application_layer/dns/Makefile
@@ -1,0 +1,5 @@
+SRC :=
+
+SUBMODULE := 1
+
+include $(RIOTBASE)/Makefile.base
